refactor: rename namespace-approval folder to package-name-approval
Completes the terminology rename started in #44931. This PR only renames
the source and test folders and updates all import paths accordingly.

No logic changes - pure folder rename for consistency with the new
package-name-* label convention.
