+++ .github/CODEOWNERS
@@ -331,8 +331,8 @@
 /.github/arm-leases/     @raosuhas @tejaswiminnu @akhilailla @msevanhi @vikeshi26 @vidapour @bubbles4096
 /.github/workflows/**/arm-*     @raosuhas @tejaswiminnu @akhilailla @msevanhi
 
-## namespace-approval
-/.github/**/namespace-approval/          @samvaity @praveenkuttappan
+## package-name-approval
+/.github/**/package-name-approval/          @samvaity @praveenkuttappan
 
 # Add owners for notifications for specific pipelines
 /eng/common/pipelines/codeowners-linter.yml
