+++ .github/workflows/package-name-approval-status.yaml
@@ -31,7 +31,7 @@ jobs:
         with:
           script: |
             const { default: postResults } =
-              await import('${{ github.workspace }}/.github/workflows/src/namespace-approval/post-results.js');
+              await import('${{ github.workspace }}/.github/workflows/src/package-name-approval/post-results.js');
             await postResults({ github, context, core });
 
       - name: Set package name approval status
@@ -39,5 +39,5 @@ jobs:
         with:
           script: |
             const { default: statusCheck } =
-              await import('${{ github.workspace }}/.github/workflows/src/namespace-approval/status-check.js');
+              await import('${{ github.workspace }}/.github/workflows/src/package-name-approval/status-check.js');
             await statusCheck({ github, context, core });
