+++ .github/workflows/package-name-approval-validation.yaml
@@ -33,7 +33,7 @@ jobs:
         with:
           script: |
             const { default: validateApproval } =
-              await import('${{ github.workspace }}/.github/workflows/src/namespace-approval/validate-approval.js');
+              await import('${{ github.workspace }}/.github/workflows/src/package-name-approval/validate-approval.js');
             await validateApproval({ github, context, core });
 
       - name: Update package name approval status
@@ -41,5 +41,5 @@ jobs:
         with:
           script: |
             const { default: statusCheck } =
-              await import('${{ github.workspace }}/.github/workflows/src/namespace-approval/status-check.js');
+              await import('${{ github.workspace }}/.github/workflows/src/package-name-approval/status-check.js');
             await statusCheck({ github, context, core });
