+++ .github/workflows/src/package-name-approval/detect-namespaces.js
